requirements.txt
python-miio
python-dotenv
fastapi
uvicorn[standard]
